Moved RegisterInstance and TechyRegister to their own package
