+++ src/main/java/tbsc/techy/api/register/TechyRegister.java
@@ -15,7 +15,10 @@
  * License along with Techy.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package tbsc.techy.api;
+package tbsc.techy.api.register;
+
+import net.minecraftforge.fml.common.registry.GameRegistry;
+import net.minecraftforge.fml.common.registry.IForgeRegistryEntry;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -23,20 +26,21 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * After registration, if a field that is annotated with this annotation is found, and the class that it is
- * pointing to has/will be registered, it assigns the instance of the registered object to the field
- * that is annotated by this annotation.
- * To know what class needs to be assigned by this field, a class object and an identifier need to be passed.
- * A class to know by the type of the registered object, and an identifier if there are more than 1 instances
- * of a class registered.
+ * This annotation should be put on any class that can be registered using {@link GameRegistry#register(IForgeRegistryEntry)},
+ * therefore any class that's implementing {@link IForgeRegistryEntry}.
+ * What it'll do is register the class to the game, and based on what the class implements, register other things.
+ * For example, if a block implements {@link net.minecraft.block.ITileEntityProvider}, it'll register the tile
+ * entity for the block.
+ *
+ * If you want to store the created instance if a field, create a public static field in the same class
+ * and annotate it with {@link RegisterInstance}. Give it the same identifier you gave here, and it'll assign the instance
+ * to that field.
  *
  * Created by tbsc on 6/23/16.
  */
-@Target(ElementType.FIELD)
+@Target(value = ElementType.CONSTRUCTOR)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface RegisterInstance {
-
-    Class register();
+public @interface TechyRegister {
 
     String identifier();
